Move LeaderBasedReplicationAdmin into ReplicationEngine
